fix(DocumentRange): Move DocumentRange to rtextarea package
DocumentRange lived in org.fife.ui.rsyntaxtextarea despite being used
extensively by the sister rtextarea package (SearchEngine,
RTextAreaHighlighter, RTextArea, SearchResult), creating an unwanted
circular dependency between the two packages. It is now
org.fife.ui.rtextarea.DocumentRange.

This is a breaking API change.
