+++ RSyntaxTextArea/src/main/java/org/fife/ui/rtextarea/DocumentRange.java
@@ -6,7 +6,7 @@
  * This library is distributed under a modified BSD license.  See the included
  * LICENSE file for details.
  */
-package org.fife.ui.rsyntaxtextarea;
+package org.fife.ui.rtextarea;
 
 
 /**
