+++ RSyntaxTextArea/src/main/java/org/fife/ui/rtextarea/SearchEngine.java
@@ -20,7 +20,6 @@ import javax.swing.JTextArea;
 import javax.swing.text.BadLocationException;
 import javax.swing.text.Caret;
 
-import org.fife.ui.rsyntaxtextarea.DocumentRange;
 import org.fife.ui.rsyntaxtextarea.RSyntaxUtilities;
