+++ RSyntaxTextArea/src/test/java/org/fife/ui/rtextarea/DocumentRangeTest.java
@@ -4,7 +4,7 @@
  * This library is distributed under a modified BSD license.  See the included
  * LICENSE file for details.
  */
-package org.fife.ui.rsyntaxtextarea;
+package org.fife.ui.rtextarea;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
